Fix .jft loading

+++ source/ham/jamulfont.h
@@ -1,6 +1,7 @@
 #ifndef JAMULFONT_H
 #define JAMULFONT_H
 
+#include <stddef.h>
 #include "jamultypes.h"
 
 class MGLDraw;
@@ -15,7 +16,7 @@ struct mfont_t
 	byte spaceSize;		// # of pixels wide to make spaces
 	byte gapSize;		// # of pixels between adjacent letters
 	byte gapHeight;		// # of pixels to descend for a carriage return
-	long dataSize;		// the size in bytes of the data of the characters themselves
+	int32_t dataSize;		// the size in bytes of the data of the characters themselves
 	byte *data;			// pointer to the character data
 	byte *chars[FONT_MAX_CHARS]; // pointers to each character's data (can't have more than FONT_MAX_CHARS)
 };
